Rename bmc module to hardwareutils
Using hardwareutils instead bmc as go module name will provide
future utility packages can be added this module to be used every
place in BMO repo.
