+++ Dockerfile
@@ -11,8 +11,8 @@ COPY apis/go.mod apis/go.mod
 COPY apis/go.sum apis/go.sum
 COPY hack/tools/go.mod hack/tools/go.mod
 COPY hack/tools/go.sum hack/tools/go.sum
-COPY pkg/bmc/go.mod pkg/bmc/go.mod
-COPY pkg/bmc/go.sum pkg/bmc/go.sum
+COPY pkg/hardwareutils/go.mod pkg/hardwareutils/go.mod
+COPY pkg/hardwareutils/go.sum pkg/hardwareutils/go.sum
 RUN go mod download
 
 COPY . .
