--- pkg/bmc/go.mod
@@ -1,5 +1,0 @@
-module github.com/metal3-io/baremetal-operator/pkg/bmc
-
-go 1.16
-
-require github.com/pkg/errors v0.9.1
