+++ pkg/hardwareutils/LICENSE
