+++ pkg/hardwareutils/bmc/errors.go
