+++ pkg/hardwareutils/bmc/idrac_virtualmedia.go
