+++ pkg/hardwareutils/bmc/redfish.go
