+++ pkg/hardwareutils/bmc/redfish_virtualmedia.go
