+++ pkg/hardwareutils/go.mod
@@ -1,0 +1,5 @@
+module github.com/metal3-io/baremetal-operator/pkg/hardwareutils
+
+go 1.16
+
+require github.com/pkg/errors v0.9.1
