+++ pkg/provisioner/ironic/raid_test.go
@@ -449,7 +449,7 @@ func TestCheckRAIDConfigure(t *testing.T) {
 
 	for _, c := range cases {
 		t.Run(c.raidInterface, func(t *testing.T) {
-			err := CheckRAIDConfigure(c.raidInterface, c.RAID)
+			err := checkRAIDConfigure(c.raidInterface, c.RAID)
 			if (err != nil) != c.expectedError {
 				t.Errorf("Got unexpected error: %v", err)
 			}
